refactor: move ConnectApiService to connect package
